services/telegram-bot/pyproject.toml
[project]
name = "chatops-telegram-bot"
version = "0.1.0"
description = "Telegram bot frontend for ChatOps AI Bridge"
requires-python = ">=3.12"
dependencies = [
    "chatops-shared",
    "python-telegram-bot[webhooks]>=21.0.0",
    "httpx>=0.27.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.2.0",
    "redis[hiredis]>=5.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/telegram_bot"]

[tool.uv.sources]
chatops-shared = { workspace = true }
